refactor: extract adapter logic for algolia params so that it can be also used in the video module adapter
